refactor(ml): convert package sources to TypeScript
