+++ packages/ml/tsconfig.json
@@ -1,0 +1,13 @@
+{
+  "extends": "../../tsconfig.packages.base.json",
+  "compilerOptions": {
+    "baseUrl": ".",
+    "rootDir": ".",
+    "paths": {
+      "@react-native-firebase/app/dist/module/internal": ["../app/dist/typescript/lib/internal"],
+      "@react-native-firebase/app": ["../app/dist/typescript/lib"]
+    }
+  },
+  "include": ["lib/**/*"],
+  "exclude": ["node_modules", "dist", "__tests__", "**/*.test.ts"]
+}
